requirements-backend.txt
databricks-sdk==0.102.0
fastapi==0.136.0
pandas
python-dotenv==1.2.2
requests
uvicorn==0.44.0
